Ep 17 * Improve Test Arrangements With Factory Classes
This was quite different since laravel now uses Factory clases, you
can't replace the create() function built into the Factory, so adding a
->withTasks(1)->create() would require using a custom create method for
just that use case...  And calling $this->signIn() in the middle of a
line is less readable anyway.  So I just added createOwnedBy for when
you want to specify the owner.